# packages/shared-contracts/openapi/cloudhelm.openapi.yaml
openapi: 3.1.0
info:
  title: CloudHelm Platform API
  description: CloudHelm 平台 API。M2 提供真实数据库驱动的项目、任务、需求、设计、审批和事件底座。
  version: 0.2.0
paths:
...
  /api/tasks/{task_id}/tool-calls:
    post:
      tags:
      - tool-calls
      summary: 创建 ToolCall 记录（M2 内部联调）
      description: 写入 ToolCall 记录；M2 不执行真实工具。
      operationId: create_tool_call_api_tasks__task_id__tool_calls_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCallCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - tool-calls
      summary: 分页读取任务 ToolCall
      description: 读取某任务下的 ToolCall 记录。
      operationId: list_tool_calls_api_tasks__task_id__tool_calls_get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 分页大小。
          default: 50
          title: Limit
        description: 分页大小。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: offset cursor。
          title: Cursor
        description: offset cursor。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ToolCallRead_'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tool-calls/{tool_call_id}:
    get:
      tags:
      - tool-calls
      summary: 读取 ToolCall
      description: 读取单个 ToolCall。
      operationId: get_tool_call_api_tool_calls__tool_call_id__get
      parameters:
      - name: tool_call_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Tool Call Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
          description: 稳定错误码。
        message:
          type: string
          title: Message
          description: 面向调用方的错误描述。
        detail:
          anyOf:
          - {}
          - type: 'null'
          title: Detail
          description: 调试细节或字段错误。
        trace_id:
          type: string
          title: Trace Id
          description: 链路追踪 ID。
      type: object
      required:
      - code
      - message
      - trace_id
      title: ErrorResponse
      description: 统一错误响应结构。
...
    PageInfo:
      properties:
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          description: 本次最多返回条数。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页游标；为空表示无更多数据。
      type: object
      required:
      - limit
      title: PageInfo
      description: '分页元数据。


        M2 使用 offset cursor 的最小实现，后续如需高并发滚动查询可升级为

        基于 `created_at,id` 的稳定游标。'
...
    PageResponse_ToolCallRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ToolCallRead'
          type: array
          title: Items
          description: 当前页数据。
        page:
          $ref: '#/components/schemas/PageInfo'
          description: 分页信息。
      type: object
      required:
      - items
      - page
      title: PageResponse[ToolCallRead]
...
    RiskLevel:
      type: string
      enum:
      - L0
      - L1
      - L2
      - L3
      - L4
      title: RiskLevel
      description: CloudHelm 工具和任务风险等级。
...
    ToolCallCreate:
      properties:
        agent_run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Agent Run Id
          description: 关联 AgentRun。
        tool_name:
          type: string
          minLength: 1
          title: Tool Name
          description: 工具名称。
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
          description: 工具风险等级。
        arguments_json:
          additionalProperties: true
          type: object
          title: Arguments Json
          description: 工具参数 JSON。
        result_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result Json
          description: 工具结果 JSON。
        status:
          $ref: '#/components/schemas/ToolCallStatus'
          description: 工具调用状态。
          default: pending
        approval_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Approval Id
          description: 关联审批请求。
      type: object
      required:
      - tool_name
      - risk_level
      title: ToolCallCreate
      description: '开发/内部联调用 ToolCall 创建请求体。


        M2 仅记录工具调用元数据和参数摘要，不执行真实工具调用。'
    ToolCallRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        task_id:
          type: string
          format: uuid
          title: Task Id
        agent_run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Agent Run Id
        tool_name:
          type: string
          title: Tool Name
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        arguments_summary:
          type: string
          title: Arguments Summary
        result_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result Json
        status:
          $ref: '#/components/schemas/ToolCallStatus'
        approval_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Approval Id
        started_at:
          type: string
          format: date-time
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
      type: object
      required:
      - id
      - task_id
      - agent_run_id
      - tool_name
      - risk_level
      - arguments_summary
      - result_json
      - status
      - approval_id
      - started_at
      - finished_at
      title: ToolCallRead
      description: 'ToolCall 响应结构。


        响应只暴露 `arguments_summary`，避免把潜在敏感参数原样展示给控制台。

        完整参数仍保存在数据库，后续审计视图可按权限读取。'
    ToolCallStatus:
      type: string
      enum:
      - pending
      - running
      - succeeded
      - failed
      - waiting_approval
      - cancelled
      title: ToolCallStatus
      description: ToolCall 状态。
servers:
- url: http://127.0.0.1:18080
  description: 本地开发平台 API